Add form validation on SignIn/Up pages and fix notifications

+++ app/Models/TeamMembers.php
@@ -20,7 +20,7 @@ class TeamMembers extends Model
 
             TeamInvites::where('to_user', $userId)->orWhere('from_user', $userId)->delete();
 
-            NotificationInvites::where('from_user', $userId)->delete();
+            NotificationInvites::where('from_user', $userId)->orWhere('user_id', $userId)->delete();
 
             if (static::where('team_id', $model->team_id)->count() >= 5) {
                 TeamInvites::where('team_id', $model->team_id)->delete();
